[config] smarter handling of completely deprecated boolean options
Summary:
What was happening in that case is that a .inferconfig file containing `{ "myflag":
false }` used to be translated by CommandLineOption as `--no-` (`--no-<long>`),
because it picked the *non-deprecated* form of the option (here `myflag` would
have to be a deprecated form of the option since its non-deprecated form is
empty `""`). Instead, pick a non-empty form of the option.

Reviewed By: jberdine

Differential Revision: D20368784

fbshipit-source-id: 8e761e684