[JavaClassName] Introduce a function to strip out anonymous class suffix from a class
Summary:
We will use it in follow up diffs.

From many perspectives, if the function belongs to an anonymous class,
it is useful to know the original user-defined class.

This function makes this distinction clear.

Thanks to ngorogiannis, whos work on refactoring `Typ.name` made this module
easy enough so we can introduce unit tests!

Reviewed By: ngorogiannis

Differential Revision: D20389311

fbshipit-source-id: 408d95660