[java] bug fix for type translation
Summary:
`JavaSplitName` is used to represent Java types (in `Procname` in particular).  The type itself is a pair of string (an optional package qualifier) and a "type name" (the quotes are there because it may contain array qualifiers).

For example `java.lang.Object[][]` should be represented as
 ```
{package=Some "java.lang"; typename="Object[][]"}
```
The constructor `make` was misused to construct instead types such as
```
{package=None; typename="java.lang.Object[][]"}`
```
This is evident when we print the return type of a `Procname` non-verbosely (the default), but we still see the package qualifier.

Obviously this is not just a pretty-printing bug, the values were themselves wrong.

The fix is to use the `of_string` constructor which will parse the package and separate it correctly. Another bug (in response to this one) had to be fixed in `Procname.is_vararg` to maintain behaviour in Nullsafe and Quandary.

Reviewed By: mityal

Differential Revision: D20394146

fbshipit-source-id: 4633902eb