[racerd] optimise add_callee_accesses
Summary:
This function is used to adapt the callee summary at a call site. It did two things for every domain element in the callee summary:
- A linear search through the list of actuals.
- For each such actual, it would (repeatedly) compute its ownership (!).

For large summaries this can be substantial. The right way is to precompute the ownership for all actuals once and then simply retrieve it (via an array).

Reviewed By: jberdine

Differential Revision: D20564447

fbshipit-source-id: 1ca3121c2