[nullsafe] Refine definition of meta issues
Summary:
This is likely not the final refinement, rather one step forward.

We classify all classes by 3 categories:
- Nullsafe and 0 issues
- can add Nullsafe and will be 0 issues
- the rest (class needs improvement)

Each class will fall into exactly one category.

Error messaging is WIP, they are not intended to be surfaced to the user
just yet.

Note how this diff uses the result of the previous refactoring.

Reviewed By: artempyanykh

Differential Revision: D20512999

fbshipit-source-id: 7f462d29d