[starvation] comply with issue log contract
Summary:
`IssueLog` is used by the file-level analysis callbacks to store reports outside error logs so as to avoid racing on spec files.  Each file should generate a single issue log which is then written to an appropriate file.  The starvation checker was breaking that contract because it ostensibly needs to write out multiple issue files when analysing a single source file.

This is unnecessary, because the existing mechanisms for deduplication ensure only one issue file needs to be written out.

The whole-program mode still needs that capability, but this is implemented outside the file-analysis callback.

Reviewed By: mityal

Differential Revision: D20736135

fbshipit-source-id: 620e5484d