[pulse] Model Core Graphics create and copy functions
Summary:
This models all the Create and Copy functions from CoreGraphics, examples in the tests.
These functions all allocate memory that needs to be manually released.

The modelling of the release functions will happen in a following diff. Until then, we have some false positives in the tests.

This check is currently in biabduction, and we aim to move it to Pulse.

Reviewed By: jvillard

Differential Revision: D20626395

fbshipit-source-id: b39eae2d9