[cost] Re-locate where to print issue with heuristics
Summary:
Currenlty the cost issue is printed at the first node of a function, which is usually the first
statment of the function.  This may give a wrong impression that the cost of the statement is
changed.

This diff re-locate where to print issues with heuristics.  Going backward from the first node
lines, it looks up a line satisfying,
1. A line should start with <fname> or should include " <fname>".
2. The <fname> found in 1 should be followed by a space, '<', '(', or end of line.

Reviewed By: jvillard

Differential Revision: D20766876

fbshipit-source-id: b4fee3180