[nullsafe] Treat signatures in the class under analysis differently
Summary:
See the comments in the code why it makes logical sense.
This diff is a step forward the state when list of type violations is
independent of the mode (and we use mode solely to decide re: whether to
report or not).

This fixes majority of cases in ModePromotions.java

Reviewed By: artempyanykh

Differential Revision: D20948656

fbshipit-source-id: 82c0d530b