[sledge] Dump perf diagnostics and replays for slow queries to stderr
Summary:
Generate output on stderr containing lines such as:
```
 solve_for_vars time:        0.105 ms         0.871 ms            79 calls
```
which indicates that the current maximal time of a single
`solve_for_vars` query is `0.105 ms` and so far the total time spent in
`solve_for_vars` over `79` queries is `0.871 ms`.

At program exit, there is also a line for each timer indicating the
max query time and final accumulated duration and number of queries:
```
 solve_for_vars time:        0.173 ms        17.242 ms          1108 calls
```

Additionally, replay sexp are dumped interactively. A query exceeding
1 sec is dumped, then one exceeding 2 sec, then 4 sec, etc.

Reviewed By: jvillard

Differential Revision: D20852160

fbshipit-source-id: 0a316891e