[sledge] Fix term ordering bug between monomials and vars
Summary:
Fix the crash in
```
(And_eq () (Var (id 10) (name v))
 (Mul (((Var (id 8) (name v)) 1) ((Var (id 9) (name v)) 1)))
 ((xs ()) (sat true) (rep ())))
```

The solver for interpreted functions relies on the solution
substitutions containing mappings from variables to interpreted
applications, and never in the reverse. When solving equations
involving polynomials, this constraint is specifically
established. But for equalities involving only monomials, it could
happen that e.g. `x` was chosen as the representative of `a × b`,
which violates this constraint.

Reviewed By: jvillard

Differential Revision: D20596422

fbshipit-source-id: 69b026f03