[sledge] Strengthen normalization of division, avoiding non-reals
Summary:
Strengthen the canonizer for division and rational constant terms. It
is important to avoid constructing `Q.t` values with 0 denominators,
as they do not represent real numbers, and the algebraic manipulation
done by the rest of the solver is incorrect in that case.

Additionally strengthen the Term representation invariants to check
that all coefficients and exponents are valid real rational numbers.

Also normalize division by an integer or rational to multiplication by
a rational.

Reviewed By: jvillard

Differential Revision: D20663964

fbshipit-source-id: 210962fe0