[racerd] remove ownership init logic for locals
Summary:
Now that only races rooted at formals or globals are reported, there is no point using the ownership domain to exclude accesses to locals, so remove the code that initialises the ownership of those.

Also, remove an accidentally quadratic use of `Map.bindings |> List.find` in `FormalMap` and simplify the analysis driver.

Reviewed By: skcho

Differential Revision: D21066855

fbshipit-source-id: 126080778