[pulse] add a cache of constants to equate them
Summary:
When encountering a constant, pulse creates an abstract value (a
variable) to represent it, and remembers that it's equal to it. The
problem is that pulse doesn't yet know how to deal with the fact that
some variables are going to be equal to each other.

This hacks around this issue in the case of constants, within the same
procedure, by remembering which constants have been assigned to which
place-holder variables, and serving those variables again when the same
constant is translated again.

Limitation: this doesn't work across procedure calls as the "constant
maps" are not saved in summaries.

Something to look out for: we don't want to make `if (p == NULL)` create
a path where `p` is invalid (we only make null invalid when we see an
assignment from 0, i.e. `p = NULL;`).

Reviewed By: ezgicicek

Differential Revision: D21089961

fbshipit-source-id: 5ebb85d0a