[reporting] Don't secretly hide `Info` issues
Summary:
It is true that `Info` issues are normally not intended for the end user
and in general should be hidden by default.
However, the current behavior - show them only if `--no-filtering` is
true - is super non-intuitive and complicates already complex reporting
logic.

Lets use the general "enable/disable" mechanism for controlling this.

Reviewed By: jvillard

Differential Revision: D21154140

fbshipit-source-id: 69e4c88e4