[pulse] Model CFBridgingRelease by removing the Allocated attribute
Summary: `CFBridgingRelease` and `__bridge_transfer` which I'll model later, transfer the memory model from manual memory ref count to ARC (automatic ref count), so to avoid false positives this needs to be modelled. We can simply remove the Allocated attribute from the state, which means we won't try to track that memory anymore.

Reviewed By: skcho

Differential Revision: D21088218

fbshipit-source-id: 3520a0d59