[nullsafe] Nested classes should inherit mode from the outer one
Summary:
From the user perspective, the current behavior is confusing.
The users intutitively expect the inner class to inherit Nullsafe mode
from the outer one.

Having a class that is Nullsafe but the inner is not is hence dangerous
and misleading.

For the sake of completeness and to support gradual strictification, we
allow the nested class to improse additional strictness. Particularly,
the inner class can be Nullsafe but the outer can be not.

A follow up to this diff will include warnings telling about redundant and wrong usages of nested annotations.

Reviewed By: artempyanykh

Differential Revision: D21228055

fbshipit-source-id: 75755ad1d