[pulse] Model __bridge_transfer
Summary:
This translates the construct `ObjCBridgedCastExpr` when the cast_kind is `OBC_BridgeTransfer`, or in syntax, the cast (`__bridge_transfer`).
This cast means that the object is passed from manual memory management to ARC, so one doesn't need to call `release` manually. It is important to model this to avoid false positives.

It translates it as a builtin that we then model in Pulse, the same way we modelled `CFBridgingRelease` which does the same thing.

The name of the builtin is `__free_cf` which is not ideal but I left it like that for compatibility with biabduction. We can change it once we remove this check from biabduction.

update-submodule: facebook-clang-plugins

Reviewed By: jvillard

Differential Revision: D21176337

fbshipit-source-id: 736ceeb9b